% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseData.R
\name{baseData_sync}
\alias{baseData_sync}
\title{基础表更新}
\usage{
baseData_sync(dms_token, erp_token, FStartDate, FEndDate)
}
\arguments{
\item{dms_token}{第二个参数}

\item{FEndDate}{}
}
\value{
两个数的和
}
\description{
基础表更新
}
\examples{
baseData_sync
}
