% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salOrder.R
\name{ordersItems_sync}
\alias{ordersItems_sync}
\title{订单明细表更新}
\usage{
ordersItems_sync(dms_token, erp_token)
}
\arguments{
\item{dms_token}{第二个参数}

\item{erp_token}{}
}
\value{
两个数的和
}
\description{
订单明细表更新
}
\examples{
ordersItems_sync
}
