% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salOrder.R
\name{salOrder_sync}
\alias{salOrder_sync}
\title{销售订单主表更新}
\usage{
salOrder_sync(dms_token, erp_token, FStartDate, FEndDate)
}
\arguments{
\item{dms_token}{第二个参数}

\item{FEndDate}{}
}
\value{
两个数的和
}
\description{
销售订单主表更新
}
\examples{
salOrder_sync
}
