% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salOrder.R
\name{salOrder_view}
\alias{salOrder_view}
\title{销售订单查询}
\usage{
salOrder_view(dms_token, FStartDate, FEndDate)
}
\arguments{
\item{dms_token}{第二个参数}

\item{FEndDate}{}
}
\value{
两个数的和
}
\description{
销售订单查询
}
\examples{
salOrder_view
}
